requirements.txt
httpx
1337x
rich
python-qbittorrent
pytest